refactor n house clean
