+++ functions/utils/admin.js
@@ -1,0 +1,11 @@
+const admin = require("firebase-admin");
+const serviceAccount = require("../serviceAccountKey.json");
+
+admin.initializeApp({
+  credential: admin.credential.cert(serviceAccount),
+  databaseUrl: "https://monkey-buzz-2019.firebaseio.com"
+});
+
+const db = admin.firestore();
+
+module.exports = { admin, db };
