+++ functions/utils/firebaseConfig.js
@@ -1,0 +1,10 @@
+
+module.exports = {
+    apiKey: process.env.APIKEY,
+    authDomain: "monkey-buzz-2019.firebaseapp.com",
+    databaseURL: process.env.DATABASEURL,
+    projectId: "monkey-buzz-2019",
+    storageBucket: "monkey-buzz-2019.appspot.com",
+    messagingSenderId: "175703200110",
+    appId: "1:175703200110:web:121b212f5e683724a9f17f"
+  };
